requirements.txt
gradio>=4.0.0
streamlit>=1.24.0
together>=0.2.0
stripe>=7.0.0
pillow>=10.0.0
python-dotenv>=1.0.0
langchain>=0.1.0
pandas==1.4.2
numpy==1.22.3
opencv-python>=4.8.0
pypdf>=3.0.0
requests>=2.31.0
aiohttp>=3.8.0
beautifulsoup4==4.9.3
scrapy==2.5.1
newsapi-python==0.2.6
transformers==4.18.0
torch==1.11.0
matplotlib==3.5.2
seaborn==0.11.2
pydantic==1.9.1
rouge-score==0.0.4
grafana-api==1.0.3
validators>=0.20.0
scikit-learn>=0.24.0
matplotlib>=3.4.0
seaborn>=0.11.0